Moved items to schemas/items
