+++ packages/gb-printer-schemas/src/index.ts
@@ -1,7 +1,7 @@
-export * from './items/Frame';
-export * from './items/FrameGroup';
-export * from './items/Image';
-export * from './items/ImageGroup';
-export * from './items/Palette';
-export * from './tools/creationDate';
-export * from './tools/safeDate';
+export * from '@/schemas/items/Frame';
+export * from '@/schemas/items/FrameGroup';
+export * from '@/schemas/items/Image';
+export * from '@/schemas/items/ImageGroup';
+export * from '@/schemas/items/Palette';
+export * from '@/tools/creationDate';
+export * from '@/tools/safeDate';
