+++ packages/gb-printer-schemas/src/schemas/items/Frame.ts
