+++ packages/gb-printer-schemas/src/schemas/items/FrameGroup.ts
