+++ packages/gb-printer-schemas/src/schemas/items/Image.ts
