+++ packages/gb-printer-schemas/src/schemas/items/ImageGroup.ts
