+++ packages/gb-printer-schemas/src/schemas/items/Palette.ts
